restructure repository to follow smallweb app conventions
